+++ adapters/cx.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for CxProlog 0.98.2 or a later version
-%  Last updated on September 9, 2024
+%  Last updated on November 4, 2024
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  SPDX-FileCopyrightText: 1998-2024 Paulo Moura <pmoura@logtalk.org>
@@ -286,6 +286,7 @@ setup_call_cleanup(Setup, Call, Cleanup) :-
 % startup flags:
 '$lgt_default_flag'(settings_file, allow).
 % lint compilation flags:
+'$lgt_default_flag'(linter, default).
 '$lgt_default_flag'(unknown_entities, warning).
 '$lgt_default_flag'(unknown_predicates, warning).
 '$lgt_default_flag'(undefined_predicates, warning).